Move the threadpool to the top level
